+++ src/main/java/org/springframework/security/data/acl/config/AclJpaRepositoriesRegistrar.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.security.data.acl;
+package org.springframework.security.data.acl.config;
 
 import java.lang.annotation.Annotation;
